+++ utils/hosp-poll.c
@@ -194,6 +194,11 @@ int main(int argc, char** argv) {
     goto close_hdev;
   }
 
+  if (hid_set_nonblocking(hosp_get_device(hosp), 1) < 0) {
+    // Not a fatal error.
+    fprintf(stderr, "hid_set_nonblocking: %ls\n", hid_error(hosp_get_device(hosp)));
+  }
+
   if (!restart || !(ret = hosp_restart(hosp))) {
     ret = hosp_poll(hosp);
   }