+++ utils/hosp-set.c
@@ -123,6 +123,11 @@ int main(int argc, char** argv) {
     goto close_hdev;
   }
 
+  if (hid_set_nonblocking(hosp_get_device(hosp), 1) < 0) {
+    // Not a fatal error.
+    fprintf(stderr, "hid_set_nonblocking: %ls\n", hid_error(hosp_get_device(hosp)));
+  }
+
   ret = hosp_set_status(hosp);
 
   if (hosp_close(hosp)) {